feat: Rework with APP and API Docker
