+++ .env.example
@@ -1,15 +1,18 @@
-# Docker variables.
-PROJECT=aisrm-lewagon
+API_PORT=8500
+API_MEMORY=512Mi
+API_TAG=dev
+
+APP_PORT=8501
+APP_MEMORY=512Mi
+APP_TAG=dev
+
+PROJECT=aisrm
 REGION=europe-west1
 ARTIFACTSREPO=aisrm
-MEMORY=512Mi
-TAG=0.1
-IMAGE_NAME=aisrm
+EMAIL=youremail@gmail.com
 
-# GCloud variables.
-EMAIL=myemail@gmail.com
-
-# Makefile variables.
+PACKAGE_NAME=aisrm
+PACKAGE_VERSION=0.0.1
 PYTHON_VENV=aisrm-env
 PYTHON_VERSION=3.10.6
 RAW_DATA_URL=https://www.kaggle.com/api/v1/datasets/download/innocentmfa/crm-sales-opportunities
