--- .github/workflows/python-ci.yml
@@ -1,49 +1,0 @@
-name: AISRM CI
-
-on:
-  push:
-    branches: [ master, main ]
-  pull_request:
-    branches: [ master, main ]
-
-jobs:
-  build-check-and-tests:
-
-    runs-on: ubuntu-latest
-    environment: dev
-    
-    env:
-      PROJECT: ${{ vars.PROJECT }}
-
-    steps:
-    # Checkout branch that triggered the event.
-    - uses: actions/checkout@v3
-
-    # Install Python.
-    - name: Set up Python 3.10
-      uses: actions/setup-python@v2
-      with:
-        python-version: "3.10"
-
-    # Install dependencies.
-    - name: Install dependencies
-      run: make requirements
-
-    # Run linters.
-    - name: Code check
-      run: make lint
-
-    # Build container
-    - name: Build Docker image
-      run: make docker_build
-    
-    - name: Run Docker container
-      run: make docker_run
-    
-    # Run tests.
-    - name: Run tests
-      run: make tests
-
-    # Kill container now
-    - name: Stop Docker Container
-      run: make docker_stop
