+++ Dockerfile.api
@@ -1,0 +1,12 @@
+FROM python:3.10-slim
+
+WORKDIR /app
+
+COPY requirements-api.txt requirements.txt
+
+RUN pip install -U pip \
+    && pip install -r requirements.txt
+
+COPY api/ ./api/
+
+CMD ["sh", "-c", "uvicorn api.run:app --host 0.0.0.0 --port ${PORT}"]
