+++ api/run.py
@@ -8,18 +8,8 @@ and basic endpoints for the CRM sales opportunities system.
 from datetime import datetime
 
 from fastapi import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI()
-
-# Allowing all middleware is optional, but good practice for dev purposes.
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["*"],  # Allows all origins
-    allow_credentials=True,
-    allow_methods=["*"],  # Allows all methods
-    allow_headers=["*"],  # Allows all headers
-)
 
 
 @app.get("/")
