+++ requirements-app.txt
@@ -1,0 +1,1 @@
+streamlit
