--- requirements-prod.txt
@@ -1,12 +1,0 @@
-# Testing
-pytest
-pylint
-ipdb
-black
-
-# Backend
-fastapi
-uvicorn
-
-# Frontend
-streamlit
